added mamba model

+++ standalone_eval/eval.py
@@ -18,6 +18,8 @@ def compute_average_precision_detection_wrapper(
 
 def compute_mr_ap(submission, ground_truth, iou_thds=np.linspace(0.5, 0.95, 10),
                   max_gt_windows=None, max_pred_windows=10, num_workers=8, chunksize=50):
+    if len(submission) == 0:
+        return {'average' : 0}
     iou_thds = [float(f"{e:.2f}") for e in iou_thds]
     pred_qid2data = defaultdict(list)
     for d in submission:
@@ -71,6 +73,8 @@ def compute_mr_ap(submission, ground_truth, iou_thds=np.linspace(0.5, 0.95, 10),
 
 def compute_mr_r1(submission, ground_truth, iou_thds=np.linspace(0.5, 0.95, 10)):
     """If a predicted segment has IoU >= iou_thd with one of the 1st GT segment, we define it positive"""
+    if len(submission) == 0:
+        return {'average' : 0}
     iou_thds = [float(f"{e:.2f}") for e in iou_thds]
     pred_qid2window = {d["qid"]: d["pred_relevant_windows"][0][:2] for d in submission}  # :2 rm scores
     # gt_qid2window = {d["qid"]: d["relevant_windows"][0] for d in ground_truth}
